Reset the counter at not exactly midnight
Changing days at midnight is a load of crap, because facebook-obsessed
millennials don't go to sleep until at least 2. So the counter resets
at 8 a.m., and if you wake up earlier than that, stop doing that.

Using the moment library instead of JSON to handle the date
serialization, because JSON was being weird with time zones or
something?